pdf_rag/requirements.txt
langchain
pypdf2
python-dotenv
streamlit
langchain-openai    
langchain-community
faiss-cpu
python-dotenv
openai